Update in progress. Still transitioning to use Room and View Models.
* In build.gradle, added location of where the Room database schema
  should be exported and added lines to allow for Java 8 code (such as
  lambda functions).

* In NacAlarmDatabase, added option to export the schema of the Room database.

* In NacAlarmDatabase, added an executor service, instead of the async
  task, because the latter is deprecated.

* In NacAlarmDatabase, added code that will handle the database
  transition from SQLite to Room, as well as inserting the first initial
  alarm, if the app was just installed. This class will store the
  application context, briefly, until the database is created and/or
  migrated. After these steps occur, or if neither occurred, the context
  will be set to null.

* Added return types to the NacAlarmDao, so that I can get the row ID of
  an inserted alarm, for instance. In order to do this, the Future type
  needed to be used.

* In NacAlarmViewModel, added checks if the alarm is null, before
  executing any actions.

* In NacMainActivity, properly handle copy/delete/restore. This entails
  being able to undo properly, and getting the alarm ID from an alarm
  that was copied, by using the Future type, mentioned above.

* In many files, changed type of alarm ID from int to long.

* Moved over the NacCardHolder code from the adapter to the main
  activity. This is now done and the respective code that was in the
  adapter, commented, is now removed.

* In NacCardHolder, renamed OnDeleteClickedListener to
  OnCardDeleteClickedListener and changed the arguments to the same as
  the other listeners: NacCardHolder and NacAlarm.

* In NacCardHolder, add OnCardMediaClickedListener in order to start the
  media activity from outside the NacCardHolder class. Namely, from the
  main activity.

* In NacCardTouchHelper, moved the PreviousSwipe class to its own file,
  and allowed it to be handled in the main activity.

* In NacCardTouchHelper, changed the name of the listener functions from
  onSwipedLeft() to onCopySwipe() and onSwipedRight() to onDeleteSwipe().

* In NacCardTouchHelper, when the card is swiped to copy, manually set
  the X-position of the card back to 0, instead of leaving it
  off-screen.

* In NacGeneralSettings, updated to use the new activity result
  structure. The request code is no longer needed.

* In NacMediaActivity, updated code to use the new onAttachFragment()
  method.

* In NacMediaFragment, updated the logic to properly set the media that
  was selected. If this was being called from the main activity, an
  alarm will be set, and the appropriate view model will be updated,
  alternatively, if it is called from General Settings, only the media
  string will be set, and will need to be passed as the result of the
  activity when it is finished.

+++ app/src/main/java/com/nfcalarmclock/NacActiveAlarmNotification.java
@@ -133,10 +133,10 @@ public class NacActiveAlarmNotification
 	{
 		Context context = this.getContext();
 		NacAlarm alarm = this.getAlarm();
-		int id = (alarm == null) ? 0 : alarm.getId();
+		long id = (alarm == null) ? 0 : alarm.getId();
 		Intent intent = NacIntent.createAlarmActivity(context, alarm);
 
-		return PendingIntent.getActivity(context, id, intent,
+		return PendingIntent.getActivity(context, (int)id, intent,
 			PendingIntent.FLAG_UPDATE_CURRENT);
 	}
